[EDITED] reduced modules

+++ src/main/java/com/shaked/service/controllers/MediaController.java
@@ -1,18 +1,12 @@
 package com.shaked.service.controllers;
 
 import com.shaked.service.excepctions.NotFoundException;
-import com.shaked.service.models.Article;
 import com.shaked.service.models.Media;
-import com.shaked.service.repositories.ArticleRepository;
 import com.shaked.service.repositories.MediaRepository;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
 
 
 @RestController
@@ -27,7 +21,7 @@ public class MediaController {
 
     @GetMapping("{id}")
     public @ResponseBody
-    ResponseEntity getMedia(@PathVariable long id) {
+    ResponseEntity getMedia(@PathVariable Integer id) {
         Media media = mediaRepository.findById(id).orElseThrow(() -> new NotFoundException(String.valueOf(id)));
         return ResponseEntity.ok()
                 .contentType(MediaType.IMAGE_JPEG)
