+++ src/main/java/com/shaked/service/repositories/ContactRepository.java
@@ -1,6 +1,5 @@
 package com.shaked.service.repositories;
 
-import com.shaked.service.models.Article;
 import com.shaked.service.models.Contact;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
@@ -8,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ContactRepository extends CrudRepository<Contact, Long> {
+public interface ContactRepository extends CrudRepository<Contact, Integer> {
 
     @Override
     List<Contact> findAll();
