+++ src/main/java/com/shaked/service/repositories/MediaRepository.java
@@ -1,12 +1,11 @@
 package com.shaked.service.repositories;
 
 import com.shaked.service.models.Media;
-import com.shaked.service.models.TextContainer;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MediaRepository extends CrudRepository<Media, Long> {
+public interface MediaRepository extends CrudRepository<Media, Integer> {
 
     List<Media> findAll();
